revert changes to NoteWithLabels

+++ app/src/main/java/com/github/androidpirate/slicknotes/data/NoteWithLabels.java
@@ -43,7 +43,7 @@ public class NoteWithLabels {
         return note;
     }
 
-    public long getNoteId() {
+    public int getNoteId() {
         return note.getNoteId();
     }
